Fix render_views for anonymous controllers on 1.8.7.
Rails uses a controller's class name to determine the layout to use when
rendering views but it does so with an inherited hook. This means that
setting the name of a class *after* it inherits is too late. This is not
an issue in Ruby 1.9.2 as the name of an anonymous module is set to nil
but in 1.8.7 it is set to the empty string (which, of course, is a
truthy value). This causes Rails to throw a "can't convert nil into
String" error when attempting to render views from an anonymous
controller.

We can fix this by declaring an anonymous controller's name *on
creation* and then evaluating the body of the controller with
`class_eval`.