Add a bit of info about Capybara::DSL to README
I personally found it a bit confusing to figure out how to use the
Capybara DSL until I read the Capybara docs at the bottom of the
README.  It wasn't immediately obvious that this doc even /existed/
when I first read the README as I wasn't interested in customising
rake tasks.  So I have made another link to the doc in the request
spec section and explained that by default the DSL is only available
in spec/features.