`subject` is defined as a method rather than with the `subject` declarative method
There is currently some odd interactions between rspec-rails and
rspec-core when it comes to `let` definitions.

If we call `subject` in an `included` hook, the dynamically generated
`LetDefinitions` module gets inserted into the ancestor chain earlier
than normal. This placement no longer allows `let` definitions to
override methods defined on other modules that are brought in via
`config.include`.

We may raise a warning in future (see discussion in #738)

[Fixes #738]

Conflicts:
	Changelog.md