Add missing Changelog entries.
Note the helpful contributions made by Rodrigo Rosenfeld Rosas and
Giovanni Cappellotto.

The removal of the `ActiveModel` dependency was added to the `Breaking
Changes` section as there is a possibility (though it should be very
small) that this could break projects. Projects which implicitly depend
on `ActiveModel` to be loaded due to rspec-rails's dependency have
incorrectly setup their Gemfile.