Move common rspec dependencies into custom Gemfile
When the example app is created for running cukes we need to setup a
Gemfile. It needs to share the same dependencies as this repo. To
facilitate this, move the common dependencies into a custom Gemfile.

Switch to using the proper bundler `eval_gemfile`, which properly
handles path location differences, instead of reading a file then using
raw `eval`.