Order the changes by importance.
Put the more important changes at the top of the page so that they are
read first. Placing the helper file changes first as the other changes
reference `rails_helper`. The next most important change is the
disabling of inferred metadata; as it has been a common source of
upgrade questions.