Extract rails gem loading into shared Gemfile.
When the example app is generated it puts a Rails version that is
defined by currently installed version. This works great in most
situtations, except `master` or other edge branches. The version listed
is the current version of Rails, so they diverge.

Move how we load rails into a custom Gemfile, which we eval ourselves.
We then remove the rails version put into the gemfile and share our
custom version. Due to when `rake` and other shell processes fork, the
RAILS_VERSION environment variable is properly passed through.