Check for ActiveRecord configurations for fixture support
Currently ActiveRecord fixtures are included if ActiveRecord is defined,
but not set up. On Rails < 4.1.0 this was not a problem, as it had a
test in ActiveRecord::TestFixtures to see if ActiveRecord was
configured. On Rails 4.1.0 this check was removed, causing rspec-rails
to always try to connect to a Database even if there is not one.