Remove warning configuration from spec_helper.rb
Resolve #1162

In RSpec 3.0, `rspec --init` added `--warnings` to `.rspec`. This wasn't
a good decision and many users asked how to get rid of the metric ton of
warnings they were getting from rails, gems, etc, so we've changed it so
that it'll be in the generated `spec_helper.rb` instead, in the
commented out section, so that users have to op-in to it.

However, many gems used by Rails apps are not warning free. Leaving this
config option turned on by default in the `spec_helper.rb` configuration
will continue to cause issues for many users.

This completely removes the setting from the generated `spec_helper.rb`,
along with the comment suggesting it is recommended to be enabled. We do
not suggest that it is enabled for Rails apps by default.