Extend rspec-mocks to see store_accessor methods
Rails has a number of ways of adding dynamic methods to models.

Verifying doubles in mocks currently fail on all dynamic methods that
aren't explicitly set in method_missing, because the model does not
recognise them as a method at the time the mock verifies.

However - many of Active Record's dynamic methods have a very fixed and
easy-to-check list... so this extension does that and adds these dynamic
methods to the allowed set of valid methods

This commit starts that ball rolling by recognising the "store_accessor"
methods (which (de)serialise columns on an Active Record)