Include bundle cache.
Unlike the other repos the rspec-rails build does not run their spec
suites (since rspec-rails is a consumer it is implicitly testing them).
Due to this we do not need the extra caching logic which is added in
rspec/rspec-dev#102.

We do however want to preserve teh bundle path. Since additional example
apps also use this path we update them as well.