Disable the test-unit gem autorunner.
For some reason the test-unit gem loads it's autorunner by default.
Though the test-unit included with Ruby does not. The original code for
loading test-unit pre-dates widespread usage of the test-unit gem.

We took care of disabling the autorunner for Ruby 2.2+ since the
test-unit gem is now required for Rails 3.2. For older Ruby versions we
attempt to only load the `test/unit/assertions` module. When using the
`test-unit` gem however, Rails will auto require it for us, which loads
the main test-unit code, thus enabling the autorunner.