Include `rspec:request` generator.
This includes a generator for "request" specs. There already exists a
generator which creates these specs, but it is called "integration".
The existing name comes from the legacy implementations attempting to
mirror the Rails test style names for the RSpec equivalent specs.

Since these two generators are identical, except for the name, this
makes the new `RequestGenerator` a subclass of the existing
`IntegrationGenerator`. Alternatively, we could have decided to make
`RequestGenerator` the parent class, moving the existing functionality
into it. While this is arguably the "proper" way of thinking about the
generators, it could potentially break some existing applications; even
though these classes are marked `private`. There's no harm for now in
keeping semver compliant by making this generator additive.

Since the two generators are exactly the same this uses a shared example
to keep the specs in-sync.

Resolve #1377