Re-write the Railtie to use a modular features
This is the first step at trying to improve the ability to support
multiple Rails versions. It extracts the majority of the logic for
`ActionMailer` configuration into clear modules.

In addition to making the code cleaner, this allows the ancestor chain
to be inspected to understand what configurations settings are included.