Switch mailer examples on rails version.
Rails 5 changes the class name generated for mailers from looking like
`Thing` to looking like `ThingMailer`. Our cukes depend on this name,
and so break as is. Here, I've added two tags, `@pre_rails_5` and
`@post_rails_5`.

The modification to the rakefile changes how we execute cukes based on
the rails version, filtering to the correct cukes.