Fix empty template rendering for Rails 4.2.5.1
Closes #1532.

Refactor the template rendering overrides in controller specs to be more
compatible with the Action View public api. Previously `EmptyTemplatePathSetDecorator`
overrode the `find_all` method to wrap a path set. However `ActionView::PathSet`
is not public api so instead create a subclass of `ActionView::FileSystemResolver`
to return an empty template and treat `view_paths` as an array of string-like objects.