Check if path is a path before decorating it
When you prepend or append view paths, you can choose to do so with either a
path (which may be looked up on the file system) or an instance of
`ActionView::Resolver` which will then know how to do the lookup. We should only
wrap paths (`String` or `Pathname` instances) in an `EmptyTemplateResolver`.