Delegate properly to custom Resolver instances
Introduce `EmptyTemplateResolverFactory` for creating instances of either
`EmptyTemplateResolverDecorator` or `EmptyTemplateFileSystemResolver`, depending
on whether the given path is an `ActionView::Resolver` instance or not.

The new `EmptyTemplateResolverDecorator` class simply delegates all method calls
to its given resolver, except for `#find_templates` which it overrides to return
templates that render with `EmptyTemplateHandler`.