In this commit: * just fixed the railtie
As of Rails 5.1.0 gems can register a directories to work with `rake
notes`

In this case we can tell Rails it to now look in the `/spec` folder when
searching for annotations like `# TODO:` `# FIXME` etc.

This was added in this PR:
rails/rails#25692

Resulting in:

```bash

$ rake notes
app/spec/models/article_spec.rb:
  * [2] [TODO] add more specs
  * [3] [FIXME] Spec A is broken
  * [4] [OPTIMIZE] improve this test

```

Thats one less daily annoyance.

John