Clarify spec_helper vs rails_helper
When greenfielding a new app & coming from older versions of rspec, the out-of-box experience was confusing for me. Since `.rspec` required `spec_helper` I figured I'd be able to simply write specs & run them.  So I was confounded when my first attempt yielded undefined constant `Rails`.

Searching for answers showed I wasn't the only one. Might as well add a short explanation here, as it's not exactly self-explanatory and even the comments in the files don't really spell it out.