Fix noisy log error
We've been seeing this error in our test logs:

    Could not log "render_template.action_view" event.
    NoMethodError: undefined method `example_group' for nil:NilClass
    .../rspec-rails-3.6.0/lib/rspec/rails/view_rendering.rb:67:in `current_example_group'
    .../rspec-rails-3.6.0/lib/rspec/rails/view_rendering.rb:71:in `render_template'
    ...

This is while running an rspec rails feature spec using capybara, so
this is a separate thread running webrick. This was also during a before
block. In any case, this should probably fail gracefully when there is
no example.

While updating this code I got an "ExampleGroup doesn't respond to
render_views?", so built in that case too.