Only configure RSpec::Mocks if it is fully loaded
It's possible for the `::RSpec::Mocks` constant to exist, but for it not
to have the `configuration` method present. This can happen if another
library has required some files from _within_ the `rspec/mocks` library,
but nothing else has fully loaded it.

For a concrete example, see https://github.com/krisleech/wisper-rspec/issues/21

While allowing external libraries to use parts of this one might not be
something that's explicitly supported, we can avoid some unexpected
exceptions by adding a little extra check to our detection of whether or
not the mocking library is loaded.