Add --no-rc flag for more reliable local testing
The rspec-rails acceptance tests use `rails new` to generate a sample
app for running tests. If a developer running the rspec-rails test suite
has a `.railsrc` file, this will affect the app that is generated,
causing specs to fail.

This commit fixes this by passing the `--no-rc` flag so that `rails new`
ignores the developer's local `.railsrc`.