Truncate more method names in system tests
ref: https://github.com/rspec/rspec-rails/pull/1999/files

If system test fails, a screenshot is taken, but if the file name exceeds 255 characters(e.g. using turnip), the following error will be output.

```
File name too long @ rb_sysopen
```

The reason is the test fails, `failures_` is appended to first of file name.
So I truncated it to a certain length so that the length of the file name does not become long.