Fix `errors` method check in be_valid matcher
If the `.be_valid?` matcher fails, it checks for the presence of a
`.errors` method and uses it, if present, in the failure message.  This
works great, unless the subject happens to include an `errors` method
that takes an argument, in which case we get an error (because we call
errors with 0 arguments).

We can fix this with a simple arity check.