Be conservative on which Rails version is supported
`rspec-rails` 3.9 can be installed with Rails 6.1, because it does not limit which rails version is supported, but it actually have issues with Rails 6.1. It is not possible to add restriction on the next point release (let's say 3.9.2), because bundler will then install previous version (3.9.1) that does not have the dependency restriction.

Moving forward with rspec-rails 4.1, I think we should be more conservative on which Rails version is supported. If it happens that the next rails version (6.2 or 7.0) is still compatible, it will be easy to make a point release of rspec-rails 4.1.x with the upper bound dependency on Rails updated.