Fix FixtureSupport's run_in_transaction? method
`ActiveRecord::TestFixture`'s `uses_transaction` is designed to be used
like this:

```ruby
uses_transaction :the_test_method_name
```

And in RSpec, the method name would be the example's name.

```ruby
uses_transaction "does someting"

it "does someting" {}
```

But in the current implementation, it's passing the example object
instead of its name, which would always fail the name comparison in
https://github.com/rails/rails/blob/adc0146a07ccc72405aec78ccb65aac3502a4300/activerecord/lib/active_record/test_fixtures.rb#L94-L97

So this commit fixes the issue by passing the example's name instead of
the example object.