Verify ActiveJob job signature matches arguments
Guards against false-positives where both the implementation
and have_enqueued_job or have_been_enqueued matchers are passed
the wrong arguments which don't match the job's signature.

Prior to this change it was possible for the implmentation and
spec to pass whilst expecting the wrong number of arguments,
or mixing up the positional/keyword arguments. This resulted
in a passing test and the implementation failing at runtime.