Clean up generator for generator specs
Don't force pluralization of the generator name in the generated spec
and use the full generator name in the spec description.

Also clean up an extra whitespace and empty line in the template.

Note that the generator class constant cannot be directly used in the
spec description as it isn't loaded by default.