Use headless driver for next Rails release
In the next release of Rails, the default driver was switched from
`:chrome` to `:headless_chrome` as see in: https://github.com/rails/rails/pull/50512
This is to ensure the new [ci template][] will "work out of the box".

However, this will not work with applications using `rspec-rails`, since
it still defaults to `:selenium`. Instead, GitHub actions will fail with
the following error:

```
Selenium::WebDriver::Error::SessionNotCreatedError:
            session not created: Chrome failed to start: exited normally.
              (session not created: DevToolsActivePort file doesn't exist)
              (The process started from chrome location /usr/bin/google-chrome is no longer running, so ChromeDriver is assuming that Chrome has crashed.)
```

[ci template]: https://github.com/rails/rails/blob/main/railties/lib/rails/generators/rails/app/templates/github/ci.yml.tt