Improve implicit description for RSpec::Rails::Matchers::ActionCable::HaveBroadcastedTo
Previously, tests defined like this:

```rb
it do
  expect { some_action }.to have_broadcasted_to("stream_name").from_channel(SomeChannel).with("some_data")
end
```

would implicitly be given the description `have broadcasted to`, which is derived from the name of the matcher method.
This didn't carry any extra information about the data we're expecting to be broadcast from the channel - behaviour
which would be very welcome in combination with Turbo Streams. With a matcher for Turbo Streams, we might benefit from
a description like "broadcast exactly 1 messages to stream with turbo-stream[action="append"][target="some_list"]".

Co-authored-by: Jon Rowe <hello@jonrowe.co.uk>
Co-authored-by: Phil Pirozhkov <pirj@users.noreply.github.com>